* Refactored code * Made share logic * Completed with unit tests of local and remote sources

+++ core/network/src/main/java/com/promovatestproject/network/okhttp/OkHttpInitializer.kt
@@ -6,7 +6,7 @@ import okhttp3.logging.HttpLoggingInterceptor
 import java.util.concurrent.TimeUnit
 
 object OkHttpInitializer {
-    private const val DEFAULT_TIMEOUT = 60
+    private const val DEFAULT_TIMEOUT = 60L
 
     fun createGeneralOkHttpClient(
         enableLogging: Boolean,
@@ -14,9 +14,9 @@ object OkHttpInitializer {
         vararg interceptors: Interceptor
     ): OkHttpClient {
         val builder = (okHttpClient?.newBuilder() ?: OkHttpClient.Builder())
-            .callTimeout(DEFAULT_TIMEOUT.toLong(), TimeUnit.SECONDS)
-            .readTimeout(DEFAULT_TIMEOUT.toLong(), TimeUnit.SECONDS)
-            .writeTimeout(DEFAULT_TIMEOUT.toLong(), TimeUnit.SECONDS)
+            .callTimeout(DEFAULT_TIMEOUT, TimeUnit.SECONDS)
+            .readTimeout(DEFAULT_TIMEOUT, TimeUnit.SECONDS)
+            .writeTimeout(DEFAULT_TIMEOUT, TimeUnit.SECONDS)
             .retryOnConnectionFailure(true)
             .apply {
                 interceptors.forEach {
